Separate interfaces and impls of breeze wrapper
