+++ common/src/main/java/edu/snu/cay/common/math/vector/breeze/SparseVector.java
@@ -16,6 +16,8 @@
 package edu.snu.cay.common.math.vector.breeze;
 
 import breeze.linalg.package$;
+import edu.snu.cay.common.math.vector.Vector;
+import edu.snu.cay.common.math.vector.VectorEntry;
 import scala.Tuple2;
 import scala.collection.JavaConversions;
