+++ dolphin/src/main/java/edu/snu/cay/dolphin/examples/ml/algorithms/regression/LinearRegMainCtrlTask.java
@@ -15,7 +15,7 @@
  */
 package edu.snu.cay.dolphin.examples.ml.algorithms.regression;
 
-import edu.snu.cay.common.math.vector.breeze.VectorFactory;
+import edu.snu.cay.common.math.vector.VectorFactory;
 import edu.snu.cay.dolphin.examples.ml.data.LinearModel;
 import edu.snu.cay.dolphin.examples.ml.parameters.Dimension;
 import edu.snu.cay.dolphin.examples.ml.parameters.MaxIterations;
