+++ dolphin/src/main/java/edu/snu/cay/dolphin/examples/ml/converge/LinearModelConvEuclidean.java
@@ -15,7 +15,7 @@
  */
 package edu.snu.cay.dolphin.examples.ml.converge;
 
-import edu.snu.cay.common.math.vector.breeze.Vector;
+import edu.snu.cay.common.math.vector.Vector;
 import edu.snu.cay.dolphin.examples.ml.data.LinearModel;
 import edu.snu.cay.dolphin.examples.ml.parameters.ConvergenceThreshold;
 import org.apache.reef.tang.annotations.Parameter;
