+++ dolphin/src/main/java/edu/snu/cay/dolphin/examples/ml/regularization/Regularization.java
@@ -15,7 +15,7 @@
  */
 package edu.snu.cay.dolphin.examples.ml.regularization;
 
-import edu.snu.cay.common.math.vector.breeze.Vector;
+import edu.snu.cay.common.math.vector.Vector;
 import edu.snu.cay.dolphin.examples.ml.data.Model;
 import org.apache.reef.tang.annotations.DefaultImplementation;
