+++ dolphin/src/main/java/edu/snu/cay/dolphin/examples/ml/sub/LinearRegSummaryCodec.java
@@ -15,8 +15,8 @@
  */
 package edu.snu.cay.dolphin.examples.ml.sub;
 
-import edu.snu.cay.common.math.vector.breeze.Vector;
-import edu.snu.cay.common.math.vector.breeze.VectorFactory;
+import edu.snu.cay.common.math.vector.Vector;
+import edu.snu.cay.common.math.vector.VectorFactory;
 import edu.snu.cay.dolphin.examples.ml.data.LinearModel;
 import edu.snu.cay.dolphin.examples.ml.data.LinearRegSummary;
 import org.apache.reef.io.serialization.Codec;
