+++ dolphin/src/main/java/edu/snu/cay/dolphin/examples/ml/sub/SparseRowCodec.java
@@ -15,7 +15,7 @@
  */
 package edu.snu.cay.dolphin.examples.ml.sub;
 
-import edu.snu.cay.common.math.vector.breeze.Vector;
+import edu.snu.cay.common.math.vector.Vector;
 import edu.snu.cay.dolphin.examples.ml.data.Row;
 import org.apache.reef.io.network.impl.StreamingCodec;
 import org.apache.reef.io.serialization.Codec;
