+++ dolphin/src/main/java/edu/snu/cay/dolphin/examples/ml/sub/SparseVectorCodec.java
@@ -15,9 +15,9 @@
  */
 package edu.snu.cay.dolphin.examples.ml.sub;
 
-import edu.snu.cay.common.math.vector.breeze.Vector;
-import edu.snu.cay.common.math.vector.breeze.VectorEntry;
-import edu.snu.cay.common.math.vector.breeze.VectorFactory;
+import edu.snu.cay.common.math.vector.Vector;
+import edu.snu.cay.common.math.vector.VectorEntry;
+import edu.snu.cay.common.math.vector.VectorFactory;
 import org.apache.reef.io.network.impl.StreamingCodec;
 import org.apache.reef.io.serialization.Codec;
