+++ dolphin/src/test/java/edu/snu/cay/dolphin/examples/ml/sub/SparseRowCodecTest.java
@@ -15,8 +15,8 @@
  */
 package edu.snu.cay.dolphin.examples.ml.sub;
 
-import edu.snu.cay.common.math.vector.breeze.Vector;
-import edu.snu.cay.common.math.vector.breeze.VectorFactory;
+import edu.snu.cay.common.math.vector.Vector;
+import edu.snu.cay.common.math.vector.VectorFactory;
 import edu.snu.cay.dolphin.examples.ml.data.Row;
 import org.apache.mahout.common.RandomUtils;
 import org.apache.reef.tang.Injector;
