+++ dolphin/src/test/java/edu/snu/cay/dolphin/examples/ml/sub/SparseVectorCodecTest.java
@@ -15,8 +15,8 @@
  */
 package edu.snu.cay.dolphin.examples.ml.sub;
 
-import edu.snu.cay.common.math.vector.breeze.Vector;
-import edu.snu.cay.common.math.vector.breeze.VectorFactory;
+import edu.snu.cay.common.math.vector.Vector;
+import edu.snu.cay.common.math.vector.VectorFactory;
 import org.apache.mahout.common.RandomUtils;
 import org.apache.reef.tang.Injector;
 import org.apache.reef.tang.Tang;
